Changed all filenames for qc -> quantum conversion, added all but gates based on mqtopt -> quantum conversion
