+++ lib/mqt-plugin.cpp
@@ -9,7 +9,7 @@
  */
 
 #include "mlir/Conversion/CatalystQuantumToQC/CatalystQuantumToQC.h"
-// #include "mlir/Conversion/QCToCatalystQuantum/QCToCatalystQuantum.h"
+#include "mlir/Conversion/QCToCatalystQuantum/QCToCatalystQuantum.h"
 #include "mlir/Dialect/QC/IR/QCDialect.h"
 
 #include <llvm/Config/llvm-config.h>
@@ -44,6 +44,6 @@ extern "C" LLVM_ATTRIBUTE_WEAK PassPluginLibraryInfo mlirGetPassPluginInfo() {
             // Note: mqt::ir::opt::registerQCPasses() is not called to avoid
             // pulling in transpilation transforms that require LLVM 21
             mqt::ir::conversions::registerCatalystQuantumToQCPasses();
-            // mqt::ir::conversions::registerQCToCatalystQuantumPasses();
+            mqt::ir::conversions::registerQCToCatalystQuantumPasses();
           }};
 }
